Refactor JS Patricia trees to reduce code duplication
