+++ packages/reputation-miner/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "@colony/colony-js-contract-loader-fs": "1.8.1",
     "bn.js": "^4.11.8",
-    "ethers": "^4.0.13",
+    "ethers": "^4.0.17",
     "express": "^4.16.3",
     "ganache-core": "2.1.6",
     "sqlite": "^3.0.0",
